+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog][changelog] and this project adheres to
 
 ## Unreleased
 
+### Removed
+
+- Support for the environment variable `CHUNKYARD_PASSCMD`
+
 ## 16.1.0 - 2022-04-29
 
 ### Added